fix(select): select popover hide bug
